separating puppeteer/electron concerns from app provider concerns
